refactor: move compoents to sub-folders
