+++ src/table/components/body/TableTotals.tsx
@@ -1,12 +1,13 @@
 import React, { memo, useMemo } from 'react';
 
-import { useContextSelector, TableContext } from '../context';
-import { getTotalsCellStyle } from '../utils/styling-utils';
-import { handleTotalKeyDown } from '../utils/handle-key-press';
-import { removeTabAndFocusCell } from '../utils/accessibility-utils';
-import { StyledHeadRow, StyledTotalsCell } from '../styles';
-import { TableTotalsProps } from '../types';
-import CellText from './CellText';
+import { useContextSelector, TableContext } from '../../context';
+import { getTotalsCellStyle } from '../../utils/styling-utils';
+import { handleTotalKeyDown } from '../../utils/handle-key-press';
+import { removeTabAndFocusCell } from '../../utils/accessibility-utils';
+import { StyledHeadRow } from '../head/styles';
+import { StyledTotalsCell } from './styles';
+import { TableTotalsProps } from '../../types';
+import CellText from '../CellText';
 
 function TableTotals({
   rootElement,
