+++ src/table/components/body/__tests__/TableTotals.spec.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import { render, fireEvent } from '@testing-library/react';
 import { stardust } from '@nebula.js/stardust';
 import TableTotals from '../TableTotals';
-import { TableContextProvider } from '../../context';
-import { getTotalPosition } from '../../../handle-data';
-import * as handleKeyPress from '../../utils/handle-key-press';
-import * as handleAccessibility from '../../utils/accessibility-utils';
-import { generateLayout } from '../../../__test__/generate-test-data';
-import { TableData, ExtendedTheme, TableLayout, ExtendedSelectionAPI } from '../../../types';
+import { TableContextProvider } from '../../../context';
+import { getTotalPosition } from '../../../../handle-data';
+import * as handleKeyPress from '../../../utils/handle-key-press';
+import * as handleAccessibility from '../../../utils/accessibility-utils';
+import { generateLayout } from '../../../../__test__/generate-test-data';
+import { TableData, ExtendedTheme, TableLayout, ExtendedSelectionAPI } from '../../../../types';
 
 describe('<TableTotals />', () => {
   const rootElement = {} as HTMLElement;
