+++ src/table/components/footer/__tests__/PaginationContent.spec.tsx
@@ -3,8 +3,8 @@ import { render, fireEvent } from '@testing-library/react';
 import { stardust } from '@nebula.js/stardust';
 
 import PaginationContent from '../PaginationContent';
-import * as handleAccessibility from '../../utils/accessibility-utils';
-import { Announce, ExtendedTheme, ExtendedTranslator, PageInfo, SetPageInfo, TableData } from '../../../types';
+import * as handleAccessibility from '../../../utils/accessibility-utils';
+import { Announce, ExtendedTheme, ExtendedTranslator, PageInfo, SetPageInfo, TableData } from '../../../../types';
 
 describe('<PaginationContent />', () => {
   let theme: ExtendedTheme;
