+++ src/table/components/virtualized-table/Cell.tsx
@@ -28,7 +28,7 @@ const Cell = ({ columnIndex, rowIndex, style, data }: CellProps) => {
           display: 'flex',
           alignItems: 'center',
           borderColor: bodyStyle.borderColor,
-          borderStyle: bodyStyle.borderStyle,
+          borderStyle: 'solid',
           borderWidth: '0px 1px 1px 0px',
           justifyContent: columns[columnIndex].align,
           boxSizing: 'border-box',
@@ -57,7 +57,7 @@ const Cell = ({ columnIndex, rowIndex, style, data }: CellProps) => {
       style={{
         ...style,
         borderColor: bodyStyle.borderColor,
-        borderStyle: bodyStyle.borderStyle,
+        borderStyle: 'solid',
         borderWidth: '0px 1px 1px 0px',
         boxSizing: 'border-box',
       }}
