+++ src/table/components/virtualized-table/Header.tsx
@@ -23,7 +23,7 @@ const Header = (props: HeaderProps) => {
         overflow: 'hidden',
         backgroundColor: headerStyle.backgroundColor,
         borderColor: headerStyle.borderColor,
-        borderStyle: headerStyle.borderStyle,
+        borderStyle: 'solid',
         borderWidth: '1px 0px',
       }}
       itemCount={layout.qHyperCube.qSize.qcx}
