+++ src/table/components/virtualized-table/Wrapper.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import { WrapperProps } from './types';
-import PaginationContent from '../PaginationContent';
-import { StyledTableWrapper } from '../../styles';
+import PaginationContent from '../footer/PaginationContent';
+import { StyledTableWrapper } from '../styles';
 import { PageInfo, TableData } from '../../../types';
-import FooterWrapper from '../FooterWrapper';
+import FooterWrapper from '../footer/FooterWrapper';
 import TableContainer from './TableContainer';
 import { MAX_PAGE_SIZE } from './constants';
