+++ src/table/utils/get-cell-renderer.ts
@@ -1,7 +1,7 @@
-import { StyledBodyCell } from '../styles';
-import withColumnStyling from '../components/withColumnStyling';
-import withSelections from '../components/withSelections';
-import withStyling from '../components/withStyling';
+import { StyledBodyCell } from '../components/body/styles';
+import withColumnStyling from '../components/body/withColumnStyling';
+import withSelections from '../components/body/withSelections';
+import withStyling from '../components/body/withStyling';
 
 export default function getCellRenderer(hasColumnStyling: boolean, isSelectionsEnabled: boolean) {
   // withStyling always runs last, applying whatever styling it gets
